container/app/requirements.txt
channels==4.1.0
cycler==0.12.1
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
mysqlclient>=1.3.13
PyMySQL==1.1.1
pynput==1.7.7
pyOpenSSL==24.1.0
pyparsing==3.1.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tqdm==4.66.4
Pillow==10.4.0